start moving code into proper locations
