+++ renderlib/gesture/gesture.h
@@ -8,140 +8,13 @@
 #include "CCamera.h"
 #include "SceneView.h"
 #include "graphics/gl/Util.h"
+#include "graphics/glsl/GLGuiShader.h"
 
-#include <chrono>
 #include <memory>
 #include <vector>
 
 class Scene;
 class RenderSettings;
-
-static const char* vertex_shader_text =
-  R"(
-    #version 400 core
-
-    layout (location = 0) in vec3 vPos;
-    layout (location = 1) in vec2 vUV;
-    layout (location = 2) in vec4 vCol;
-    layout (location = 3) in uint vCode;
-
-    uniform mat4 projection;
-    uniform int picking;
-
-    out vec4 Frag_color;
-    out vec2 Frag_UV;
-
-    void main()
-    {
-        Frag_UV = vUV;
-        if (picking == 1) {
-          Frag_color = vec4(float(vCode & 0xffu) / 255.0,
-                            float((vCode >> 8) & 0xffu) / 255.0,
-                            float((vCode >> 16) & 0xffu) / 255.0,
-                            1.0);
-        }
-        else {
-          Frag_color = vCol;
-        }
-        gl_Position = projection * vec4(vPos, 1.0);
-    }
-    )";
-
-static const char* fragment_shader_text =
-  R"(
-    #version 400 core
-    in vec4 Frag_color;
-    in vec2 Frag_UV;
-    in vec4 gl_FragCoord;
-    uniform int picking;  //< draw for display or for picking? Picking has no texture.
-    uniform sampler2D Texture;
-    out vec4 outputF;
-
-    void main()
-    {
-        vec4 result = Frag_color;
-
-        // When drawing selection codes, everything is opaque.
-        if (picking == 1)
-            result.w = 1.0;
-
-        // Gesture geometry handshake: any uv value below -64 means
-        // no texture lookup. Check VertsCode::k_noTexture
-        if (picking == 0 && Frag_UV.s > -64)
-            result *= texture(Texture, Frag_UV.st);
-
-        // Gesture geometry handshake: any uv equal to -128 means
-        // overlay a checkerboard pattern. Check VertsCode::k_marqueePattern
-        if (Frag_UV.s == -128)
-        {
-            // Create a pixel checkerboard pattern used for marquee
-            // selection
-            int x = int(gl_FragCoord.x); int y = int(gl_FragCoord.y);
-            if (((x+y) & 1) == 0) result = vec4(0,0,0,1);
-        }
-        outputF = result;
-    }
-    )";
-class GLGuiShader : public GLShaderProgram
-{
-public:
-  GLGuiShader()
-    : GLShaderProgram()
-  {
-    utilMakeSimpleProgram(vertex_shader_text, fragment_shader_text);
-
-    m_loc_proj = uniformLocation("projection");
-    m_loc_vpos = attributeLocation("vPos");
-    m_loc_vuv = attributeLocation("vUV");
-    m_loc_vcol = attributeLocation("vCol");
-    m_loc_vcode = attributeLocation("vCode");
-  }
-
-  ~GLGuiShader() {}
-
-  void configure(bool display, GLuint textureId)
-  {
-    bind();
-    check_gl("bind gesture draw shader");
-
-    glEnableVertexAttribArray(m_loc_vpos);
-    check_gl("enable vertex attrib array 0");
-    glEnableVertexAttribArray(m_loc_vuv);
-    check_gl("enable vertex attrib array 1");
-    glEnableVertexAttribArray(m_loc_vcol);
-    check_gl("enable vertex attrib array 2");
-    glEnableVertexAttribArray(m_loc_vcode);
-    check_gl("enable vertex attrib array 3");
-
-    glUniform1i(uniformLocation("picking"), display ? 0 : 1);
-    check_gl("set picking uniform");
-    if (display)
-      glUniform1i(uniformLocation("Texture"), 0);
-    else
-      glUniform1i(uniformLocation("Texture"), 1);
-    check_gl("set texture uniform");
-    glActiveTexture(GL_TEXTURE0);
-    glBindTexture(GL_TEXTURE_2D, textureId);
-    check_gl("bind texture");
-  }
-  void cleanup()
-  {
-    release();
-    glDisableVertexAttribArray(m_loc_vpos);
-    glDisableVertexAttribArray(m_loc_vuv);
-    glDisableVertexAttribArray(m_loc_vcol);
-    glDisableVertexAttribArray(m_loc_vcode);
-
-    glActiveTexture(GL_TEXTURE0);
-    glBindTexture(GL_TEXTURE_2D, 0);
-  }
-
-  int m_loc_proj;
-  int m_loc_vpos;
-  int m_loc_vuv;
-  int m_loc_vcol;
-  int m_loc_vcode;
-};
 
 struct Shaders
 {
@@ -516,95 +389,3 @@ struct Gesture
   };
   Graphics graphics;
 };
-
-// on main thread, have one clock instance that calls tick on every iteration of main event loop
-// then call mainwindow.gesture.setTimeIncrement(clock.timeIncrement);
-struct Clock
-{
-  Clock()
-    : time(0)
-    , timeIncrement(0)
-  {
-    time = Clock::now();
-  }
-
-  static double now()
-  {
-    auto currentDateTime = std::chrono::system_clock::now();
-    const auto ms =
-      std::chrono::time_point_cast<std::chrono::milliseconds>(currentDateTime).time_since_epoch().count() % 1000;
-    return ms / 1000.0;
-  }
-
-  double tick()
-  {
-    double currentTime = Clock::now();
-
-    timeIncrement = currentTime - time;
-    time = currentTime;
-    return timeIncrement;
-  }
-
-  double time;
-  double timeIncrement;
-};
-
-#if 0
-// The swap interval indicates how many frames to wait until swapping the buffers, commonly
-    // known as vsync. By default, the swap interval is zero, meaning buffer swapping will occur
-    // immediately. On fast machines, many of those frames will never be seen, as the screen is
-    // still only updated typically 60-75 times per second, so this wastes a lot of CPU and GPU
-    // cycles.
-    // Also, because the buffers will be swapped in the middle the screen update, leading to screen
-    // tearing. For these reasons, applications will typically want to set the swap interval to
-    // one. It can be set to higher values, but this is usually not recommended, because of the
-    // input latency it leads to.
-    glfwSwapInterval(0);
-
-    double lastTimeCheck = 0;
-    double increments = 0;
-    int frameRate = 0;
-
-    // Enter app event loop. The loop will continue until we quit the program.
-    // At each iteration with update the app state and redraw the window.
-    while (!glfwWindowShouldClose(mainWindow.handle))
-    {
-        // Advance clock, we need to know how much time passed since last iteration in order
-        // to advance any animated interaction we may have in our editor.
-        clock.tick();
-        mainWindow.gesture.setTimeIncrement(clock.timeIncrement);
-
-        // Display frame rate in window title
-        float interval = clock.time - lastTimeCheck; //< Interval in seconds
-        increments += 1;
-        if (interval >= 1.0)
-        {
-            // Compute average frame rate over the last second, if different than what we
-            // display previously, update the window title.
-            int newFrameRate = roundf(increments / interval);
-            if (frameRate != newFrameRate)
-            {
-                frameRate = newFrameRate;
-
-                char title[256];
-                snprintf(title, 256, "%s | %d fps", windowTitle, frameRate);
-                glfwSetWindowTitle(mainWindow.handle, title);
-            }
-            lastTimeCheck = clock.time;
-            increments = 0;
-        }
-
-        // Keep running
-        mainWindow.Render(engine, clock);
-
-        // Make sure we consumed any unused input event before we poll new events.
-        mainWindow.gesture.input.consume();
-
-        // Three ways to control event processing
-        // - glfwPollEvents(); dispatches any event in the queue and returns right away
-        // - glfwWaitEvents(); does as the name says, no processing or draw until the user interacts
-        // - glfwWaitEventsTimeout(0.7); somewhere in between, maybe to limit too tight spinning?
-        glfwPollEvents();
-    }
-
-#endif
