+++ renderlib/graphics/glsl/CMakeLists.txt
@@ -6,6 +6,8 @@ target_sources(renderlib PRIVATE
 	"${CMAKE_CURRENT_SOURCE_DIR}/GLBasicVolumeShader.h"	
 	"${CMAKE_CURRENT_SOURCE_DIR}/GLCopyShader.cpp"
 	"${CMAKE_CURRENT_SOURCE_DIR}/GLCopyShader.h"	
+	"${CMAKE_CURRENT_SOURCE_DIR}/GLGuiShader.cpp"
+	"${CMAKE_CURRENT_SOURCE_DIR}/GLGuiShader.h"	
 	"${CMAKE_CURRENT_SOURCE_DIR}/GLPTVolumeShader.cpp"
 	"${CMAKE_CURRENT_SOURCE_DIR}/GLPTVolumeShader.h"	
 	"${CMAKE_CURRENT_SOURCE_DIR}/GLToneMapShader.cpp"
